+++ relay-general/src/processor/traits.rs
@@ -91,6 +91,7 @@ pub trait Processor: Sized {
     process_method!(process_span, crate::protocol::Span);
     process_method!(process_trace_context, crate::protocol::TraceContext);
     process_method!(process_native_image_path, crate::protocol::NativeImagePath);
+    process_method!(process_contexts, crate::protocol::Contexts);
 
     fn process_other(
         &mut self,